refactor(runtime): move agent status control into its own module
`world.rs` reached 1205 non-test lines against the workspace's 1200-line
structure rule, between the provider-pool work on main and the held-inference
replay here.

`agent_status`, `set_status`, `pause`, `resume`, `replay_held_inference` and
`cancel` move to `world/control.rs`. That is one coherent unit rather than an
arbitrary cut to get under a number: every method reads or writes
`AgentState::status`, they all reach it through `agent_status`'s foreign-id
guard, and the guards they apply to one another - pause only from
`Active`/`Idle`, resume only from `Paused`/`Idle` - only make sense read side
by side.

No behaviour change; the methods are inherent on `PipelineWorld` either way.
